Moves everything into pages folder.
