+++ src/pages/EventsPage/index.js
