+++ src/pages/PointsPage/index.js
