+++ src/pages/PointsPage/point_display.js
