+++ src/pages/PointsPage/points.js
